Fix rendering of particles (partially)
It renders particles on center. Meaning, the abs position of a particle
is the center of the image being drawn. I can confirm this works for
the attack and move indicator. However, testing this for siege/tank shooting
this seems not to work. I believe it has to do with the placements of particles there.